Extract methods in separate classes
